requirements.txt
aiogram==3.4.1
python-dotenv==1.0.0
pyautogui==0.9.53
pywin32==306
keyboard==0.13.5
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0
aiohttp==3.9.1
pyperclip==1.8.2
numpy<2
google-generativeai==0.3.0
